refactor esdb event store
